pyproject.toml
[tool.poetry]
name = "hastalavistamoney"
version = "1.4.0"
description = "Home Accounting Project"
authors = ["Alexander Pavlov <dev@pavlovteam.ru>"]
packages = [
    { include = "hasta_la_vista_money" },
]

[tool.poetry.dependencies]
python = "^3.11.2"
python-dotenv = "^1.0.0"
dj-database-url = "^2.1.0"
psycopg2-binary = "^2.9.3"
whitenoise = "^6.0.0"
django-filter = "^23.3"
PyYAML = "^6.0"
Django = "^4.1.7"
urllib3 = "^2.0.2"
django-csp = "^3.7"
channels = "^4.0.0"
Pillow = "^10.0.1"
numpy = "^1.24.3"
django-axes = "^6.0.5"
Twisted = {extras = ["http2", "tls"], version = "^24.7.0"}
python-dateutil = "^2.8.2"
django-debug-toolbar = "^4.2.0"
daphne = "^4.0.0"
sentry-sdk = "^1.37.1"
django-rosetta = "^0.10.0"
crispy-bootstrap5 = "^2024.2"
django-crispy-forms = "^2.1"
djangorestframework = "^3.15.0"

[tool.poetry.dev-dependencies]
Faker = "^13.7.0"
coverage = "^7.3.2"
myst-parser = "^2.0.0"
black = "^24.3.0"
flower = "^2.0.1"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.4.0"
yesqa = "^1.5.0"
mypy = "^1.5.1"
bump2version = "^1.0.1"
icecream = "^2.1.3"
werkzeug = "^3.0.1"
django-extensions = "^3.2.3"
ruff = "^0.6.2"
mkdocs = "^1.6.1"
mkdocstrings = {extras = ["python"], version = "^0.26.1"}
mkdocs-material = "^9.5.39"
django-stubs = {extras = ["compatible-mypy"], version = "^5.1.0"}
types-python-dateutil = "^2.9.0.20241003"

[build-system]
requires = ["poetry-core>=1.4.0"]
build-backend = "poetry.core.masonry.api"

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
docstring-code-format = true
exclude = [
    "migrations",
    ".venv",
    ".git"
]

[mypy]
module = 'django.*'
ignore_missing_imports = true

[tool.mypy]
plugins = ['mypy_django_plugin.main']

[tool.django-stubs]
django_settings_module = 'config.django.base'

[tool.black]
line-length = 80
skip-string-normalization = true
exclude = '''
(
  /(
    | \.git
    | \.mypy_cache
    | \.venv
    | _build
    | build
    | dist
    | migrations/.*
    /(
      | migrations
    )/
    /(
        | bootstrap
    )
  )/
)
'''
